update table key

+++ database/migrations/2019_07_24_090328_create_product_attributes_table.php
@@ -14,10 +14,10 @@ class CreateProductAttributesTable extends Migration
     public function up()
     {
         Schema::create('product_attributes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('quantity');
+            $table->bigIncrements('id');
+            $table->bigInteger('quantity');
             $table->decimal('price')->nullable();
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products');
             $table->timestamps();
         });
